Moved adapter to a separate file
Issue #5
